Prepare move of safe_embedding_lookup_sparse from framework/ to layers/ Change: 126346230
